重新编写使用例子，带关联查询和分页查询

+++ Easy4net/Common/TypeUtils.cs
@@ -69,7 +69,7 @@ namespace Easy4net.Common
                     break;
                 case "System.Single":
                     if (!isNullOrEmpty(value))
-                        value = Convert.ToDouble(value);
+                        value = Convert.ToSingle(value);
                     break;
                 case "System.Decimal":
                     if (!isNullOrEmpty(value))
